+++ storage/driver/minio/contract_test.go
@@ -4,9 +4,9 @@ import (
 	"testing"
 
 	"github.com/morehao/golib/storage"
-	"github.com/morehao/golib/storage/driver/s3driver"
+	"github.com/morehao/golib/storage/driver/s3base"
 )
 
 func TestContract(t *testing.T) {
-	var _ storage.Storage = (*s3driver.Driver)(nil)
+	var _ storage.Storage = (*s3base.Driver)(nil)
 }
