+++ storage/driver/oss/driver.go
@@ -2,7 +2,7 @@ package oss
 
 import (
 	"github.com/morehao/golib/storage"
-	"github.com/morehao/golib/storage/driver/s3driver"
+	"github.com/morehao/golib/storage/driver/s3base"
 )
 
 func init() {
@@ -10,7 +10,7 @@ func init() {
 	storage.RegisterPathBuilder(string(storage.DriverOSS), NewPathBuilder)
 }
 
-var _ storage.Storage = (*s3driver.Driver)(nil)
+var _ storage.Storage = (*s3base.Driver)(nil)
 
 func NewPathBuilder(cfg storage.Config) storage.PathBuilder {
 	return &storage.S3PathBuilder{
@@ -22,5 +22,5 @@ func NewPathBuilder(cfg storage.Config) storage.PathBuilder {
 }
 
 func New(cfg storage.Config) (storage.Storage, error) {
-	return s3driver.New(cfg, NewPathBuilder(cfg))
+	return s3base.New(cfg, NewPathBuilder(cfg))
 }
